model_api_server/requirements.txt
uvicorn
starlette
sse-starlette
python-multipart
pyyaml
requests
# torch
# transformers
# accelerate
# sentencepiece
# openai
opencc